fix: treat 'Group not found' as retryable routing error for topology changes
When topology changes (adding/killing nodes), clients can connect to nodes
that don't have certain space groups. The 'Group not found' error was not
being treated as retryable, causing operations to fail immediately.

Changes:
- Add peGroupNotFound error kind in protocol/types.nim
- Update isNotLeaderError() to return true for 'Group not found' messages
- Update isNotLeader() to check for peGroupNotFound error kind
- Update server.nim PUT/DELETE handlers to send sendNotLeaderError for
  'Group not found' errors, triggering client retry on different node
- Add refreshClientMetadata calls in test_space_multinode.nim after
  adding node 6 and after killing node 5

This enables client-side retry logic to properly handle topology changes
by refreshing metadata and retrying on a node that has the group.

+++ src/fractio/storage/mvcc/garbage_collector.nim
@@ -317,7 +317,7 @@ proc collectVersionsForTransaction*(gc: GarbageCollector,
   ## Useful after transaction commit/abort
 
   gc.logger.info("Collecting versions for transaction",
-    {"transactionId": $int64(transactionId)}.toTable)
+    {"transactionId": $transactionId}.toTable)
 
   result = GCResult(
     success: false,
@@ -355,7 +355,7 @@ proc collectVersionsForTransaction*(gc: GarbageCollector,
             bytesCollected += encodedKey.len + valueResult.get().len
 
             gc.logger.debug("Collected intent for transaction",
-              {"transactionId": $int64(transactionId),
+              {"transactionId": $transactionId,
                "key": mvccKey.userKey}.toTable)
 
       discard scanIter.next()
@@ -365,14 +365,14 @@ proc collectVersionsForTransaction*(gc: GarbageCollector,
     result.success = true
 
     gc.logger.info("Collected versions for transaction",
-      {"transactionId": $int64(transactionId),
+      {"transactionId": $transactionId,
        "versionsCollected": $versionsCollected,
        "bytesCollected": $bytesCollected}.toTable)
 
   except Exception as e:
     result.error = "Failed to collect versions for transaction: " & e.msg
     gc.logger.error("Failed to collect versions for transaction",
-      {"transactionId": $int64(transactionId),
+      {"transactionId": $transactionId,
        "error": e.msg}.toTable)
 
 proc startGC*(gc: GarbageCollector) =
